requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.1
schedule==1.2.1
pytz==2024.1
python-dateutil==2.8.2
click==8.1.7 